Навёл красоту (#11)
* Добавлена регистрация, идентификация и аутентификация

* Добавлен префикс у названий таблиц, захешированы пароли у "начальных" юзеров

* Поиск поста по имени.

* Счетчик лайков.

*Вывод самых популярных постов на главной.

*Header -> Title

* Добавлен Id для UserDetails, аутентификация по логину без учета регистра

* Поиск по имени без учета регистра.

* Сортировка по дате и по лайкам.

* Допилил регистрацию

* Добавли тему для Bootstrap

* Украшаем

* Украшаем дальше

* Добавлена обработка ошибок (без эксепшенов)

* Украшаем 2

* Пагинация.

* Поправила шаблоны.

* Поравила вывод последних статей.

* Украшаем дальше 2

* Украшаем 3

* Добавила нормальные данные в БД.

* Украсил

+++ src/main/java/ru/bahusdivus/bhope/controllers/PostsController.java
@@ -1,21 +1,23 @@
 package ru.bahusdivus.bhope.controllers;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.data.domain.Page;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.ui.Model;
-import ru.bahusdivus.bhope.dto.PostDto;
 import ru.bahusdivus.bhope.dto.CommentDto;
+import ru.bahusdivus.bhope.dto.LikeAjaxRequest;
+import ru.bahusdivus.bhope.dto.PostDto;
 import ru.bahusdivus.bhope.dto.PostWithCommentsDto;
 import ru.bahusdivus.bhope.dto.UserDto;
 import ru.bahusdivus.bhope.services.CommentsService;
 import ru.bahusdivus.bhope.services.PostsService;
 import ru.bahusdivus.bhope.services.UserService;
+import ru.bahusdivus.bhope.utils.UserDetailsUserImpl;
 
+import java.io.IOException;
 import java.util.List;
 
 @Controller
@@ -34,7 +36,8 @@ public class PostsController {
     @RequestMapping("/")
     public String getIndex(@AuthenticationPrincipal UserDetails userDetails, Model model) {
         model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
-        List<PostDto> posts = postsService.getPostsByLike();
+        model.addAttribute("userDetails", userDetails);
+        List<PostDto> posts = postsService.getPostsByDateByLike();
         UserDto userDto = new UserDto();
         PostDto postDto = new PostDto();
         model.addAttribute("postDto", postDto);
@@ -43,86 +46,125 @@ public class PostsController {
         return "index";
     }
 
-    @RequestMapping("/posts")
-    public String getPosts(@AuthenticationPrincipal UserDetails userDetails, Model model) {
+    @RequestMapping("/posts/byDate/{pageNumber}")
+    public String getPosts(@PathVariable("pageNumber") int pageNumber,
+                           @AuthenticationPrincipal UserDetails userDetails,
+                           Model model) {
         model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
-        List<PostDto> posts = postsService.getPostsOrderByDate();
+        model.addAttribute("userDetails", userDetails);
+        Page<PostDto> posts = postsService.getPostsOrderByDate(pageNumber);
         UserDto userDto = new UserDto();
         PostDto postDto = new PostDto();
+        String byType = "byDate";
+        model.addAttribute("byType", byType);
         model.addAttribute("postDto", postDto);
         model.addAttribute("user", userDto);
         model.addAttribute("posts", posts);
         return "posts";
     }
 
-    @RequestMapping("/posts/byLike")
-    public String getPostsByLike(@AuthenticationPrincipal UserDetails userDetails, Model model) {
+    @RequestMapping("/posts/byLike/{pageNumber}")
+    public String getPostsByLike(@PathVariable("pageNumber") int pageNumber,
+                                 @AuthenticationPrincipal UserDetails userDetails,
+                                 Model model) {
         model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
-        List<PostDto> posts = postsService.getPostsOrderByLikeCount();
+        model.addAttribute("userDetails", userDetails);
+        Page<PostDto> posts = postsService.getPostsOrderByLikeCount(pageNumber);
         UserDto userDto = new UserDto();
         PostDto postDto = new PostDto();
+        String byType = "byLike";
+        model.addAttribute("byType", byType);
         model.addAttribute("postDto", postDto);
         model.addAttribute("user", userDto);
         model.addAttribute("posts", posts);
         return "posts";
     }
 
-    @RequestMapping("/users/{userId}/posts")
+    @RequestMapping("/users/{userId}/posts/{pageNumber}")
     public String getPostByUser(@PathVariable("userId") long userId,
+                                @PathVariable("pageNumber") int pageNumber,
                                 @AuthenticationPrincipal UserDetails userDetails,
                                 Model model) {
         model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
-        List<PostDto> posts = postsService.getPostsByUserId(userId);
+        model.addAttribute("userDetails", userDetails);
+        UserDto userDto = userService.findById(userId);
+        PostDto postDto = new PostDto();
+        Page<PostDto> posts = postsService.getPostsByUserId(userId, pageNumber);
+        model.addAttribute("postDto", postDto);
+        model.addAttribute("user", userDto);
         model.addAttribute("posts", posts);
         return "postByUser";
     }
 
-    @RequestMapping("/find/{name}")
+    @RequestMapping(value = "/find/{name}/{pageNumber}")
     public String getPostByUserName(@PathVariable("name") String name,
+                                    @PathVariable("pageNumber") int pageNumber,
                                     @AuthenticationPrincipal UserDetails userDetails,
                                     Model model) {
         model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
-        List<PostDto> posts = postsService.getPostsByUserName(name);
-        model.addAttribute("name", name);
+        model.addAttribute("userDetails", userDetails);
+        Page<PostDto> posts = postsService.getPostsByUserName(name, pageNumber);
+        UserDto userDto = new UserDto();
+        PostDto postDto = new PostDto();
+        model.addAttribute("user", userDto);
+        model.addAttribute("postDto", postDto);
+        model.addAttribute("name",name);
         model.addAttribute("posts", posts);
         return "findByUserName";
     }
 
     @RequestMapping("/post")
-    public String getNewPostForm(@AuthenticationPrincipal UserDetails userDetails,
-                                 Model model) throws IllegalAccessException {
+    public String getNewPostForm(@AuthenticationPrincipal UserDetailsUserImpl userDetails,
+                                 Model model) {
 
         if (userDetails == null) {
-            throw new IllegalAccessException("нужно зарегистрироваться");
+            model.addAttribute("errorMessage", "Нужно авторизоваться");
+            model.addAttribute("statusCode", "401");
+            return "error";
         }
-        model.addAttribute("login", userDetails.getUsername());
+        model.addAttribute("userDetails", userDetails);
 
         PostDto postDto = new PostDto();
-        postDto.setUser(new UserDto(1));
+        postDto.setUser(new UserDto(userDetails.getId()));
         model.addAttribute("post", postDto);
         return "editPost";
     }
 
     @RequestMapping("/post/{postId}/edit")
     public String getEditPostForm(@PathVariable("postId") long postId,
-                                  @AuthenticationPrincipal UserDetails userDetails,
-                                  Model model) throws IllegalAccessException {
+                                  @AuthenticationPrincipal UserDetailsUserImpl userDetails,
+                                  Model model) {
 
         if (userDetails == null) {
-            throw new IllegalAccessException("нужно зарегистрироваться");
+            model.addAttribute("errorMessage", "Нужно авторизоваться");
+            model.addAttribute("statusCode", "401");
+            return "error";
         }
-        model.addAttribute("login", userDetails.getUsername());
+        model.addAttribute("userDetails", userDetails);
 
         PostDto postDto = postsService.getPost(postId);
+        if (userDetails.getId() != postDto.getUser().getId()) {
+            model.addAttribute("errorMessage", "Нет доступа");
+            model.addAttribute("statusCode", "403");
+            return "error";
+        }
         model.addAttribute("post", postDto);
         return "editPost";
     }
 
     @RequestMapping("/post/{postId}/delete")
     public String deletePost(@PathVariable("postId") long postId,
-                             @AuthenticationPrincipal UserDetails userDetails) throws IllegalAccessException {
+                             @AuthenticationPrincipal UserDetailsUserImpl userDetails, Model model) {
         if (userDetails == null) {
-            throw new IllegalAccessException("нужно зарегистрироваться");
+            model.addAttribute("errorMessage", "Нужно авторизоваться");
+            model.addAttribute("statusCode", "401");
+            return "error";
+        }
+        if (userDetails.getId() != postsService.getPost(postId).getUser().getId()
+                && !userDetails.getAdmin()) {
+            model.addAttribute("errorMessage", "Нет доступа");
+            model.addAttribute("statusCode", "403");
+            return "error";
         }
         postsService.deletePost(postId);
         return "redirect:/";
@@ -130,50 +172,57 @@ public class PostsController {
 
     @RequestMapping(value = "savePost", method = RequestMethod.POST)
     public String savePost(@ModelAttribute PostDto postDto) {
-        postsService.savePost(postDto);
-        return "redirect:/";
+        return "redirect:/post/" + postsService.savePost(postDto).getId();
     }
 
     @RequestMapping(value = "findByUserName", method = RequestMethod.POST)
     public String findByUserName(@ModelAttribute UserDto userDto) {
-        return "redirect:/find/" + userDto.getName();
+        return "redirect:/find/" + userDto.getName() + "/0";
     }
 
     @RequestMapping(value = "incrementLikeCount", method = RequestMethod.POST)
-    public String incrementLikeCount(@ModelAttribute PostDto postDto) {
-        postsService.incrementLikeCount(postDto.getId());
-        return "redirect:/";
-    }
-
-    @RequestMapping("/post/{postId}/comment")
-    public String getNewCommentForm(@PathVariable("postId") long postId,
-                                    @AuthenticationPrincipal UserDetails userDetails,
-                                    Model model) throws IllegalAccessException {
-        return getNewCommentFormWithParent(postId, 0, userDetails, model);
+    @ResponseBody
+    public int incrementLikeCount(@RequestBody String json) throws IOException {
+        ObjectMapper objectMapper = new ObjectMapper();
+        LikeAjaxRequest request = objectMapper.readValue(json, LikeAjaxRequest.class);
+        return postsService.incrementLikeCount(request.getId());
     }
 
     @RequestMapping("/post/{postId}")
     public String getPage(@PathVariable("postId") long postId,
-                          @AuthenticationPrincipal UserDetails userDetails,
+                          @AuthenticationPrincipal UserDetailsUserImpl userDetails,
                           Model model) {
-        model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
+        model.addAttribute("userDetails", userDetails);
 
-        model.addAttribute("login", userDetails != null ? userDetails.getUsername() : null);
         PostWithCommentsDto post = commentsService.getPostWithComments(postId);
         model.addAttribute("post", post.getPost());
         model.addAttribute("comments", post.getComments());
         return "postPage";
     }
 
+    @RequestMapping("/post/{postId}/comment")
+    public String getNewCommentForm(@PathVariable("postId") long postId,
+                                    @AuthenticationPrincipal UserDetailsUserImpl userDetails,
+                                    Model model) {
+        return getNewCommentFormWithParent(postId, 0, userDetails, model);
+    }
+
     @RequestMapping("/post/{postId}/comment/{id}")
     public String getEditCommentForm(@PathVariable("postId") long postId,
                                      @PathVariable("id") long id,
-                                     @AuthenticationPrincipal UserDetails userDetails,
-                                     Model model) throws IllegalAccessException {
+                                     @AuthenticationPrincipal UserDetailsUserImpl userDetails,
+                                     Model model) {
         if (userDetails == null) {
-            throw new IllegalAccessException("нужно зарегистрироваться");
+            model.addAttribute("errorMessage", "Нужно авторизоваться");
+            model.addAttribute("statusCode", "401");
+            return "error";
         }
         CommentDto comment = commentsService.getComment(id);
+        if (userDetails.getId() != comment.getUser().getId()) {
+            model.addAttribute("errorMessage", "Нет доступа");
+            model.addAttribute("statusCode", "403");
+            return "error";
+        }
         model.addAttribute("comment", comment);
         return "commentPage";
     }
@@ -181,18 +230,20 @@ public class PostsController {
     @RequestMapping("/post/{postId}/comment/{id}/comment")
     public String getNewCommentFormWithParent(@PathVariable("postId") long postId,
                                               @PathVariable("id") long id,
-                                              @AuthenticationPrincipal UserDetails userDetails,
-                                              Model model) throws IllegalAccessException {
+                                              @AuthenticationPrincipal UserDetailsUserImpl userDetails,
+                                              Model model) {
 
         if (userDetails == null) {
-            throw new IllegalAccessException("нужно зарегистрироваться");
+            model.addAttribute("errorMessage", "Нужно авторизоваться");
+            model.addAttribute("statusCode", "401");
+            return "error";
         }
-        model.addAttribute("login", userDetails.getUsername());
+        model.addAttribute("userDetails", userDetails);
 
         CommentDto commentDto = new CommentDto();
         commentDto.setPost(postId);
         commentDto.setParent(id);
-        long userId = userService.findByLogin(userDetails.getUsername()).getId();
+        long userId = userDetails.getId();
         commentDto.setUser(new UserDto(userId));
         model.addAttribute("comment", commentDto);
         return "commentPage";
@@ -201,9 +252,17 @@ public class PostsController {
     @RequestMapping("/post/{postId}/comment/{id}/delete")
     public String deleteComment(@PathVariable("postId") long postId,
                                 @PathVariable("id") long id,
-                                @AuthenticationPrincipal UserDetails userDetails) throws IllegalAccessException {
+                                @AuthenticationPrincipal UserDetailsUserImpl userDetails, Model model) {
         if (userDetails == null) {
-            throw new IllegalAccessException("нужно зарегистрироваться");
+            model.addAttribute("errorMessage", "Нужно авторизоваться");
+            model.addAttribute("statusCode", "401");
+            return "error";
+        }
+        if (userDetails.getId() != commentsService.getComment(id).getUser().getId()
+                && !userDetails.getAdmin()) {
+            model.addAttribute("errorMessage", "Нет доступа");
+            model.addAttribute("statusCode", "403");
+            return "error";
         }
         commentsService.deleteComment(id);
         return "redirect:/post/" + postId;
